[FIX] borrando videos con el mismo nombre

+++ video_2_720.sh
@@ -8,6 +8,7 @@ done
 for video in *.mp4; do
     # Obtener la resolución del vídeo utilizando ffprobe
     resolucion=$(ffprobe -v error -select_streams v:0 -show_entries stream=width,height -of csv=p=0 "$video")
+    echo "$video $resolucion"
 
     # Extraer el ancho y el alto de la resolución
     ancho=$(echo "$resolucion" | cut -d ',' -f1)
@@ -35,19 +36,23 @@ for video in *.mp4; do
         #    rm "$video";
         # fi
         
-        if ffmpeg -y -i "$video" -threads 16 -c:v libx264 -preset veryslow -b:v 1500k -pass 1 -vf "scale=$nueva_anchura:$nueva_altura" -f mp4 /dev/null && \
-           ffmpeg    -i "$video" -threads 16 -c:v libx264 -preset veryslow -b:v 1500k -pass 2 -vf "scale=$nueva_anchura:$nueva_altura" -c:a aac "${video%.*}_720p.mp4"; then
-            rm "$video";
+        if [[ -f "${video%.*}_720p.mp4" ]]; then
+            echo "  El vídeo $video saltado"
+        else
+            if ffmpeg -y -i "$video" -threads 16 -c:v libx264 -preset veryslow -b:v 1500k -pass 1 -vf "scale=$nueva_anchura:$nueva_altura" -f mp4 /dev/null && \
+               ffmpeg    -i "$video" -threads 16 -c:v libx264 -preset veryslow -b:v 1500k -pass 2 -vf "scale=$nueva_anchura:$nueva_altura" -c:a aac "${video%.*}_720p.mp4"; then
+                rm "$video";
+            fi
+            echo "  El vídeo $video a 720p. $ancho x $alto => $nueva_anchura x $nueva_altura"
         fi
-        echo "El vídeo $video a 720p. $ancho x $alto => $nueva_anchura x $nueva_altura"
     else
         if [[ ! $video =~ _720p\.mp4$ ]]; then
             if ffmpeg -y -i "$video" -threads 16 -c:v libx264 -preset veryslow -b:v 1500k -pass 1 -f mp4 /dev/null && \
                ffmpeg    -i "$video" -threads 16 -c:v libx264 -preset veryslow -b:v 1500k -pass 2 -c:a aac "${video%.*}_720p.mp4"; then
                 rm "$video";
             fi
+            echo "  El vídeo $video crf $ancho x $alto"
         fi
-        echo "El vídeo $video crf $ancho x $alto"
     fi
     rm -f ffmpeg2pass-0.log ffmpeg2pass-0.log.mbtree;
 done
